Added a config file and config parser
In a bid to make tweaking the settings a bit more user friendly, the old
parameters.py was removed and all of the settings were collected into to
a config.ini file.

Also implemented a config parser in configure.py which handles all of
the importing of settings.

Added a few new features like "sticky", which makes partcles add
themselves to a the aggregate whenever adjacent, rather than only when
they attempted to move into an aggregate-containing space.

Added a "self_aggregate" settings, where two particles that collide will
add themselves to the aggregate.